+++ backend/Library.Application/GenreUseCases/Commands/DeleteGenreHandler.cs
@@ -13,6 +13,9 @@ namespace Library.Application.GenreUseCases.Commands
         {
             var genre = await _unitOfWork.GenreRepository.GetByIdAsync(request.Id, cancellationToken);
 
+            if (genre == null)
+                throw new NotFoundException(request.Id.ToString());
+
             _unitOfWork.GenreRepository.Delete(genre);
             await _unitOfWork.SaveChangesAsync();
         }
